+++ src/main/java/com/decrease/repositories/PaymentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Interface que define um repositório para operações de acesso a dados relacionados a pagamentos.
  */
-public interface PaymentRepository extends JpaRepository<Payment, Integer> {
+public interface PaymentRepository extends JpaRepository<Payment, Long> {
 
 }
